+++ src/math/Ray.h
@@ -5,8 +5,8 @@
 #include "Vec3.h"
 
 struct Ray3f {
-	glm::vec3 origin;
-	glm::vec3 direction;
+	glm::dvec3 origin;
+	glm::dvec3 direction;
 };
 
 class Ray {
